added `app:nginx` argument to launch the nginx webserver
We install nginx and add the argument `app:nginx` to start the nginx
server in the container. This means that nginx and owncloud containers
can be launched on physically separate docker hosts.

Since we no longer need to expose the OWNCLOUD_INSTALL_DIR volume, the
application can be setup on kubernetes.